Move defaults out of utils
